python/benchmarks/buffered_ingestion/requirements.txt
python-dotenv
sift-stack-py
ruff
pytest
pytest-benchmark
